completed initial version of SoundingDataset
SoundingDataset now has some reasonable derived quantities and
a gridded product. Its vertical axis is "iz" rather than "z" to
reflect the fact that altitude changes with time, but this may not
be the best way of doing things...

get_dataset can also now get a SoundingDataset with data_type 'sounding'
or 'soundings'.